core: Move mpi_resort_particles()
