+++ src/python/espressomd/cellsystem.pxd
@@ -26,7 +26,6 @@ cdef extern from "communication.hpp":
     void mpi_bcast_cell_structure(int cs)
     void mpi_set_use_verlet_lists(bool use_verlet_lists)
     int n_nodes
-    vector[int] mpi_resort_particles(int global_flag)
 
 cdef extern from "cells.hpp":
     int CELL_STRUCTURE_DOMDEC
@@ -41,6 +40,7 @@ cdef extern from "cells.hpp":
     const DomainDecomposition * get_domain_decomposition()
 
     vector[pair[int, int]] mpi_get_pairs(double distance)
+    vector[int] mpi_resort_particles(int global_flag)
 
 cdef extern from "tuning.hpp":
     cdef void c_tune_skin "tune_skin" (double min_skin, double max_skin, double tol, int int_steps, bool adjust_max_skin)
